+++ change-log.md
@@ -14,7 +14,8 @@
 - add: ContextInitializerConfigurations for controlling which VariableLists are applied 
 - add: ListContainsSource
 - add: PropertySource and PropertyResolver support classes, and support for object.property string expressions
-- add: SystemVariableSource and $$varname expression, SystemVariableProvider and dynamic implementation lookup 
+- add: SystemVariableSource and $$varname expression, SystemVariableProvider and dynamic implementation lookup
+- add: environment and test system variables
 
 ## 0.3